Slight change to interface to agree with streambuf

+++ src/runtime/runtime.h
@@ -142,33 +142,19 @@ class Runtime : public Asynchronous {
     //
     // In contrast to the system task and program logic interfaces, these
     // methods are scheduled immediately, regardless of the execution state of
-    // the simulation. From a design point of view, this interface is a thin
-    // wrapper around streambufs.
+    // the runtime. From a design point of view, this interface is a thin
+    // wrapper around streambufs. Attempting to perform an operation on an
+    // unrecognized or closed stream id will result in undefined behavior.
     // 
-    // Establishes an entry in the stream table and returns a unique id
-    // associated with that entry. If no file named path exists, it is created
-    // in the current working directory. The stream is opened in read/append
-    // mode and the read/write pointers are set to the 0'th byte in the file.
+    // Creates an entry in the stream table by calling new filebuf(path, in|out).
     SId fopen(const std::string& path);
-    // Removes an entry from the stream table. Does nothing if there is no
-    // stream associated with id.
+    // Removes an entry from the stream table. 
     void close(SId id);
-    // Updates the read/write pointer (r == true/false) in the stream
-    // associated with id. Does nothing if there is no stream associated with
-    // id.  Seeks past the beginning or end of a stream are clamped.
+    // streambuf operators
     void seekoff(SId id, int n, bool r);
-    // Copies up to m bytes beginning from the read pointer in the stream
-    // associated with id and advances the pointer by m bytes. m may be less
-    // than n if the end of file is encountered first.
     size_t sgetn(SId id, char* c, size_t n);
-    // Copies n characters from c starting from the write pointer in the stream
-    // associated with id. Does nothing if there is no stream associated with
-    // id.
     void sputn(SId id, const char* c, size_t n);
-    // Returns an estimate of the number of characters remanining between the
-    // read pointer and the end of the stream. Postiive result: at least this
-    // many characters. -1: end of file. 0: No information available.
-    size_t in_avail(SId id);
+    int in_avail(SId id);
 
     // Profiling Interface:
     //
